+++ src/Framework/ClassResolver/AbstractClassResolver.php
@@ -6,7 +6,7 @@ namespace Gacela\Framework\ClassResolver;
 
 use Gacela\Framework\AbstractConfig;
 use Gacela\Framework\AbstractFactory;
-use Gacela\Framework\ClassResolver\Cache\GacelaCache;
+use Gacela\Framework\ClassResolver\Cache\GacelaFileCache;
 use Gacela\Framework\ClassResolver\ClassNameFinder\ClassNameFinderInterface;
 use Gacela\Framework\ClassResolver\Config\ConfigResolver;
 use Gacela\Framework\ClassResolver\Factory\FactoryResolver;
@@ -101,7 +101,7 @@ abstract class AbstractClassResolver
     {
         if ($this->classNameFinder === null) {
             $this->classNameFinder = (new ClassResolverFactory(
-                new GacelaCache(Config::getInstance()),
+                new GacelaFileCache(Config::getInstance()),
                 new GacelaProfiler(Config::getInstance()),
                 Config::getInstance()->getSetupGacela()
             ))->createClassNameFinder();