+++ src/Framework/ClassResolver/Cache/CustomServicesPhpCache.php
@@ -8,16 +8,6 @@ final class CustomServicesPhpCache extends AbstractFileCache
 {
     public const FILENAME = 'gacela-custom-services.php';
 
-    /**
-     * @internal
-     *
-     * @return array<string,string>
-     */
-    public static function all(): array
-    {
-        return self::$cache;
-    }
-
     protected function getCacheFilename(): string
     {
         return self::FILENAME;