+++ tests/Feature/Framework/CustomCacheDirectorycustom/cache-dir/gacela-class-name.php
@@ -1,0 +1,6 @@
+<?php
+
+declare(strict_types=1);
+return [
+  '\\GacelaTest\\Feature\\Framework\\CustomCacheDirectory\\Module\\Persistence\\FakeRepository' => '\\GacelaTest\\Feature\\Framework\\CustomCacheDirectory\\Module\\Persistence\\FakeRepository',
+];