+++ tests/Feature/Framework/GacelaProfiler/Module/Persistence/FakeRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Feature\Framework\CustomNoCacheDirectory\Module\Persistence;
+namespace GacelaTest\Feature\Framework\GacelaProfiler\Module\Persistence;
 
 final class FakeRepository
 {