+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-### Unreleased
+### 0.28.0
+#### 2022-10-27
 
 - Add file cache for resolved classes.
 - Remove profiler, because it does the same as the file cache system under the hood.